refactor(AnimatePresence): extract container logic to usePresenceContainer hook
- Create use-presence-container.ts hook with all container state management
- Simplify AnimatePresence.vue from 250+ lines to 44 lines
- Remove unused useAnimatePresence function from presence.ts
